// src/core/api/generated/spring/schemas/TenantSetupRequestDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export interface TenantSetupRequestDTO {
  /**
   * @minLength 0
   * @maxLength 255
   */
  tenantName: string;
  timezone?: string;
  currency?: string;
  language?: string;
  createSampleData?: boolean;
  industry?: string;
}