// src/core/api/generated/spring/schemas/TenantInfoDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export interface TenantInfoDTO {
  tenantId?: string;
  tenantName?: string;
  active?: boolean;
  createdDate?: string;
  schemaName?: string;
  bootstrapVersion?: string;
  setupRequired?: boolean;
  setupCompleted?: boolean;
  timezone?: string;
  currency?: string;
  language?: string;
}