// src/core/api/generated/spring/schemas/TenantSetupProgressDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type TenantSetupProgressDTOStatus = typeof TenantSetupProgressDTOStatus[keyof typeof TenantSetupProgressDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TenantSetupProgressDTOStatus = {
  NOT_STARTED: 'NOT_STARTED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
} as const;